delete original main activity, improve user accessibility
