+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".GiveMainActivity" />
         <activity android:name=".DetailViewActivity" /> <!-- 카카오 로그인 시 웹 뷰를 띄우기 위한 코드 -->
         <activity
             android:name="com.kakao.auth.authorization.authcode.KakaoWebViewActivity"
