+++ app/src/main/java/com/example/bboba/ChatActivity.kt
@@ -2,7 +2,6 @@ package com.example.bboba
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.recyclerview.widget.RecyclerView
 
 class ChatActivity : AppCompatActivity() {
