+++ app/src/main/java/com/example/bboba/DetailViewActivity.kt
@@ -4,16 +4,12 @@ import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
-import android.os.Looper
 import android.util.Log
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.FragmentTransaction
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.FirebaseDatabase
@@ -23,7 +19,6 @@ import com.kakao.usermgmt.UserManagement
 import com.kakao.usermgmt.callback.MeV2ResponseCallback
 import com.kakao.usermgmt.response.MeV2Response
 import kotlinx.android.synthetic.main.activity_detail_view.*
-import kotlinx.android.synthetic.main.fragment_list.*
 import java.util.*
 
 class DetailViewActivity: AppCompatActivity() {
@@ -139,7 +134,7 @@ class DetailViewActivity: AppCompatActivity() {
                     }, 10000)//10초  (디버깅 편하게 일단 10초로 설정)
                     //val ft: FragmentTransaction = supportFragmentManager.beginTransaction()
                     //화면전환
-                    val nextIntent = Intent(context, GiveMainActivity::class.java)
+                    val nextIntent = Intent(context, MainActivity::class.java)
                     nextIntent.putExtra("fragmentNumber", fragmentNumber)
                     nextIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP) // 뒤로가기 버튼 눌렀을때 액티비티 스택에 쌓여있는 전의 화면을 불러오는데 이를 없애서 뒤로가기를 계속 눌렀을 때 중복 화면을 없앤다.
                     startActivity(nextIntent)
