+++ app/src/main/java/com/example/bboba/LoginActivity.kt
@@ -20,7 +20,7 @@ import com.kakao.usermgmt.callback.MeV2ResponseCallback
 import com.kakao.usermgmt.response.MeV2Response
 import com.kakao.util.exception.KakaoException
 import com.kakao.util.helper.Utility.getPackageInfo
-import kotlinx.android.synthetic.main.activity_main.*
+import kotlinx.android.synthetic.main.activity_main2.*
 import java.security.MessageDigest
 import java.security.NoSuchAlgorithmException
