+++ app/src/main/java/com/example/bboba/MainActivity.kt
@@ -3,7 +3,7 @@ package com.example.bboba
 import android.content.Context
 import android.content.DialogInterface
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
+import android.net.Uri
 import android.os.Bundle
 import android.util.Log
 import android.view.Menu
@@ -12,12 +12,13 @@ import android.view.View
 import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.GravityCompat
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import kotlinx.android.synthetic.main.activity_main.*
 import com.google.android.material.navigation.NavigationView
 import com.google.android.material.snackbar.Snackbar
+import com.google.android.material.tabs.TabLayout
 import com.kakao.auth.ApiResponseCallback
 import com.kakao.auth.network.response.AccessTokenInfoResponse
 import com.kakao.network.ErrorResult
@@ -27,15 +28,20 @@ import com.kakao.usermgmt.callback.MeV2ResponseCallback
 import com.kakao.usermgmt.callback.UnLinkResponseCallback
 import com.kakao.usermgmt.response.MeV2Response
 import com.kakao.util.helper.log.Logger
+import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.main_nav_header.*
 
-class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
+
+open class MainActivity : AppCompatActivity(),MapFragment.OnFragmentInteractionListener, NavigationView.OnNavigationItemSelectedListener {
+    val listFragment = ListFragment() // 하나의 프레그먼트를 계속 이용하기 위해 생성자로 생성후 변수에 저장
+    val mapFragment = MapFragment()
     val context: Context = this
 
     //카카오 계정에 로그인 되어있는지를 검사하여 로그인 정보가 없으면 바로 로그인 페이지로 이동
     //토큰이 있으면 메인액티비티가 나온다
     fun requestAccessTokenInfo(context: Context) { //context를 매개변수에 넣어줘야한다
-        com.kakao.auth.AuthService.getInstance().requestAccessTokenInfo(object:ApiResponseCallback<AccessTokenInfoResponse>() {
+        com.kakao.auth.AuthService.getInstance().requestAccessTokenInfo(object:
+            ApiResponseCallback<AccessTokenInfoResponse>() {
             override fun onSessionClosed(errorResult: ErrorResult?) {
                 val nextIntent = Intent(context, LoginActivity::class.java)
                 startActivity(nextIntent)
@@ -55,40 +61,38 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         })
     }
 
+    override fun onFragmentInteraction(uri: Uri) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
         requestAccessTokenInfo(this)//카카오 로그인 여부 검사 후 인텐트로 화면 전환
 
 
-        //액티비티 이동
-        bt_req.setOnClickListener {
-            //요청하기
-            val nextIntent = Intent(this, RequestActivity::class.java)
-            startActivity(nextIntent)
-        }
-        bt_give.setOnClickListener {
-            //제공하기
-            val nextIntent = Intent(this, GiveMainActivity::class.java)
-            startActivity(nextIntent)
-        }
+        val adapter = ViewPagerAdapter(supportFragmentManager)
+        adapter.addFragment(listFragment, "시간순 보기")
+        adapter.addFragment(mapFragment, "위치로 보기")
+        viewPager.adapter = adapter
+        tabs.setupWithViewPager(viewPager)
 
         //툴바 사용 설정
         setSupportActionBar(my_toolbar)
         // 2. 툴바 왼쪽 버튼 설정
-        supportActionBar!!.setDisplayHomeAsUpEnabled(true) // 왼쪽 버튼 사용 여부 true
-        supportActionBar!!.setHomeAsUpIndicator(R.drawable.ic_menu_black_24dp) // 왼쪽 버튼 아이콘 설정
+        supportActionBar!!.setDisplayHomeAsUpEnabled(true) // 네비게이션드로어 메뉴 사용
+        supportActionBar!!.setHomeAsUpIndicator(R.drawable.ic_menu_black_24dp) // 네비게이션드로어 메뉴 아이콘 설정
         supportActionBar!!.setDisplayShowTitleEnabled(false) // 타이틀 안보이게 하기
+
         //네비게이션뷰 리스너 등록
         navigationView.setNavigationItemSelectedListener(this)
     }
 
     // 3.툴바 메뉴 버튼을 설정
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
         menuInflater.inflate(R.menu.main_menu, menu) // menu/main_menu 메뉴를 toolbar 메뉴 버튼으로 설정
         return true
     }
+
     //툴바 메뉴 버튼이 클릭 됐을 때 콜백
     override fun onOptionsItemSelected(item: MenuItem?): Boolean {
         // 클릭된 메뉴 아이템의 아이디 마다 when 구절로 클릭시 동작을 설정한다.
@@ -130,7 +134,7 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
                 startActivity(intent)
             }
             R.id.logout -> {
-                UserManagement.getInstance().requestLogout(object:LogoutResponseCallback(){
+                UserManagement.getInstance().requestLogout(object: LogoutResponseCallback(){
                     override fun onCompleteLogout() {
                         //로그아웃 메세지 띄우기
                         val snackbar: Snackbar = Snackbar.make(my_toolbar, "로그아웃되었습니다", Snackbar.LENGTH_SHORT)
@@ -145,16 +149,11 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
                 })
             }
-//            R.id.bug_report -> { //임시로 로그인으로 사용
-//                val nextIntent = Intent(this, LoginActivity::class.java)
-//                startActivity(nextIntent)
-//            }
-            R.id.setting -> { //회원탈퇴
-                //chat test
+            R.id.chat -> {
                 val nextintent = Intent(context,ChatActivity::class.java)
                 startActivity(nextintent)
-
-                /*
+            }
+            R.id.setting -> { //회원탈퇴
                 val appendMessage: String = "회원탈퇴 하시겠습니까?"
                 AlertDialog.Builder(this).setMessage(appendMessage)
                     .setPositiveButton(getString(R.string.com_kakao_ok_button),
@@ -168,7 +167,6 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
                                         val nextIntent = Intent(context, RequestActivity::class.java)
                                         startActivity(nextIntent)
                                     }
-
                                     override fun onNotSignedUp() {
                                         val nextIntent = Intent(context, RequestActivity::class.java)
                                         startActivity(nextIntent)
@@ -188,7 +186,7 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
                             override fun onClick(dialog: DialogInterface?, which: Int) {
                                 dialog!!.dismiss()
                             }
-                        }).show()*/
+                        }).show()
             }
         }
         drawerLayout.closeDrawer(GravityCompat.START)
@@ -207,4 +205,17 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
             finishAffinity()
         }
     }
+
+    override fun onResume() {
+        super.onResume()
+        //교수님이 말씀하신 매칭 선택 후 뒤로가기를 통해 map화면으로 오는 것이 아닌 자동으로 map 화면 띄우기 구현
+
+        var fragmentnumber = this.intent.getIntExtra("fragmentNumber", 1)
+        if(fragmentnumber==2) {
+            val tabhost: TabLayout = this.findViewById(R.id.tabs)
+            tabhost.getTabAt(1)!!.select()
+        }
+    }
+
+
 }
