+++ app/src/main/java/com/example/bboba/MyGivingFragment.kt
@@ -12,9 +12,7 @@ import com.kakao.network.ErrorResult
 import com.kakao.usermgmt.UserManagement
 import com.kakao.usermgmt.callback.MeV2ResponseCallback
 import com.kakao.usermgmt.response.MeV2Response
-import kotlinx.android.synthetic.main.fragment_list.*
 import kotlinx.android.synthetic.main.fragment_my_giving.*
-import kotlinx.android.synthetic.main.fragment_my_request.*
 
 /**
  * A simple [Fragment] subclass.
