+++ app/src/main/res/menu/main_navigation_menu.xml
@@ -6,6 +6,11 @@
             android:icon="@drawable/ic_shopping_cart_black_24dp"
             android:title="나의 거래목록"
             />
+        <item
+            android:id="@+id/chat"
+            android:icon="@drawable/ic_message_black_24dp"
+            android:title="채팅"
+            />
     </group>
     <group android:checkableBehavior="single">
         <item
